project-5-hello-langchain/requirements.txt
langchain
langcahin-google-genai
python-dotenv